+++ src/test/java/io/kontur/eventapi/pdc/normalization/HpSrvSearchNormalizerTest.java
@@ -12,6 +12,7 @@ import org.wololo.geojson.FeatureCollection;
 import org.wololo.geojson.Point;
 
 import java.io.IOException;
+import java.util.List;
 import java.util.UUID;
 
 import static io.kontur.eventapi.pdc.normalization.PdcHazardNormalizer.HAZARD_PROPERTIES;
@@ -48,7 +49,8 @@ class HpSrvSearchNormalizerTest {
         assertEquals(1590675541253L, obs.getEndedAt().toInstant().toEpochMilli());
         assertEquals(1590590813468L, obs.getSourceUpdatedAt().toInstant().toEpochMilli());
         assertEquals(dataLake.getLoadedAt(), obs.getLoadedAt());
-        assertEquals("https://hazardbrief.pdc.org/TEST/ui/index.html?uuid=d26f0681-70e2-48b2-83eb-c8b9d8ef69fe", obs.getSourceUri());
+        assertEquals(List.of("https://hazardbrief.pdc.org/TEST/ui/index.html?uuid=d26f0681-70e2-48b2-83eb-c8b9d8ef69fe"),
+                obs.getSourceUri());
 
         assertNotNull(obs.getGeometries());
         checkGeometriesValue(obs.getGeometries());